+++ src/executors/threads/WorkerThread.cpp
@@ -24,7 +24,6 @@
 #include "hardware/device/AcceleratorStream.hpp"
 #include "scheduling/Scheduler.hpp"
 #include "system/If0Task.hpp"
-#include "system/PollingAPI.hpp"
 #include "system/TrackingPoints.hpp"
 #include "tasks/LoopGenerator.hpp"
 #include "tasks/Task.hpp"
@@ -83,7 +82,7 @@ void WorkerThread::body()
 		// There should not be any pre-assigned task
 		assert(_task == nullptr);
 
-		_task = Scheduler::getReadyTask(cpu);
+		_task = Scheduler::getReadyTask(cpu, this);
 		if (_task != nullptr) {
 			WorkerThread *assignedThread = _task->getThread();
 
@@ -121,9 +120,6 @@ void WorkerThread::body()
 			}
 			CPUManager::checkIfMustReturnCPU(this);
 		} else {
-			// Execute polling services
-			PollingAPI::handleServices();
-
 			// If no task is available, the CPUManager may want to idle this CPU
 			CPUManager::executeCPUManagerPolicy(cpu, IDLE_CANDIDATE);
 		}
@@ -168,6 +164,7 @@ void WorkerThread::handleTask(CPU *cpu)
 	// Execute the task
 	if (_task != nullptr) {
 		executeTask(cpu);
+
 		_task = nullptr;
 	}
 }