+++ src/scheduling/schedulers/SyncScheduler.cpp
@@ -1,12 +1,13 @@
 /*
 	This file is part of Nanos6 and is licensed under the terms contained in the COPYING file.
 
-	Copyright (C) 2019-2020 Barcelona Supercomputing Center (BSC)
+	Copyright (C) 2019-2021 Barcelona Supercomputing Center (BSC)
 */
 
 #include "SyncScheduler.hpp"
 
 #include <InstrumentScheduler.hpp>
+
 
 Task *SyncScheduler::getTask(ComputePlace *computePlace)
 {
@@ -36,17 +37,18 @@ Task *SyncScheduler::getTask(ComputePlace *computePlace)
 	// serving tasks to the rest of active compute places, except when there is work for
 	// all compute places. A compute place should stay inside the scheduler to check for
 	// deadline tasks but also for progressively resuming idle compute places when there
-	// is available work. However, device scheduler do not work like that because they
-	// already implement their progress engine using polling services. Also, external
-	// or compute places being disabled should not serve tasks for a long time
+	// is available work. However, device schedulers do not work like that because they
+	// already implement their progress engine using polling tasks. Also, external or
+	// compute places being disabled should not serve tasks for a long time
 	do {
-		size_t servedTasks = 0;
+		size_t servingIters = 0;
+		bool hasIncompatibleWork;
 
 		// Move ready tasks from add queues to the unsynchronized scheduler
 		processReadyTasks();
 
 		// Serve the rest of computes places that are waiting
-		while (servedTasks < _maxServedTasks && !_lock.empty()) {
+		while (servingIters < _maxServingIters && !_lock.empty()) {
 			// Get the index of the waiting compute place
 			computePlaceIdx = _lock.front();
 			assert(computePlaceIdx < _totalComputePlaces);
@@ -55,27 +57,35 @@ Task *SyncScheduler::getTask(ComputePlace *computePlace)
 			assert(waitingComputePlace != nullptr);
 
 			// Try to get a ready task from the scheduler
-			task = _scheduler->getReadyTask(waitingComputePlace);
+			task = _scheduler->getReadyTask(waitingComputePlace, hasIncompatibleWork);
 
 			if (task != nullptr)
 				Instrument::schedulerLockServesTask(task->getInstrumentationTaskId());
 
-			// Assign the task to the waiting compute place even if it nullptr. The
-			// responsible for serving tasks is the current compute place, and we
-			// want to avoid changing the responsible constantly, as happened in the
-			// original implementation
-			_lock.setItem(computePlaceIdx, task);
+			// If we are using the hybrid/busy policy, avoid assigning tasks even if
+			// none are found, so that threads do not spin in their body to avoid
+			// contention in here. The "responsible" thread will be the one busy
+			// iterating until the criteria of max busy iterations is met
+			if (task != nullptr || _currentBusyIters++ >= _numBusyIters || hasIncompatibleWork) {
+				// Assign the task to the waiting compute place even if it is nullptr. The
+				// responsible for serving tasks is the current compute place, and we want
+				// to avoid changing the responsible constantly, as happened in the original
+				// implementation
+				_lock.setItem(computePlaceIdx, task);
 
-			// Unblock the served compute place and advance to the next one
-			_lock.popFront();
+				// Unblock the served compute place and advance to the next one
+				_lock.popFront();
 
-			servedTasks++;
+				_currentBusyIters = 0;
+			}
+
+			servingIters++;
 			if (task == nullptr)
 				break;
 		}
 
 		// No more compute places waiting; try to get work for myself
-		task = _scheduler->getReadyTask(computePlace);
+		task = _scheduler->getReadyTask(computePlace, hasIncompatibleWork);
 
 		// Keep serving while there is no work for the current compute
 		// place or it is external/disabling